# Stop allocating a new vector for each unwind call