Up until now, each time we would unwind the native stack we would start with an empty vector, and then reserve enough space for it to hold the largest stack we had seen so far on the current thread. Instead of starting from scratch each time, save this vector in a thread-local variable so that we can reuse it from one unwind call to the next.

Note that our `RecursionGuard` prevents us from ever trying to unwind a second time on a thread before we've finished recording the first unwound stack. Note also that we don't record stack traces for deallocations, only allocations, so it won't be a problem that our deallocation hook will be called for this thread-local vector once the thread dies.