@pablogsal 2023-01-13
## Before

```
❯ hyperfine 'python -m memray run --trace-python-allocators --native -o /dev/null -m test test_list'
Benchmark 1: python -m memray run --trace-python-allocators --native -o /dev/null -m test test_list
  Time (mean ± σ):      3.517 s ±  0.035 s    [User: 3.414 s, System: 0.085 s]
  Range (min … max):    3.470 s …  3.602 s    10 runs
```

## After

```
❯ hyperfine 'python -m memray run --trace-python-allocators --native -o /dev/null -m test test_list'
Benchmark 1: python -m memray run --trace-python-allocators --native -o /dev/null -m test test_list
  Time (mean ± σ):      2.987 s ±  0.032 s    [User: 2.899 s, System: 0.061 s]
  Range (min … max):    2.948 s …  3.057 s    10 runs
```

So ~18% speedup!