@pablogsal 2023-01-13
> Note also that we don't record stack traces for deallocations, only allocations, so it won't be a problem that our deallocation hook will be called for this thread-local vector once the thread dies.

I am not sure this works. A deallocation can allocate. Indeed, this happens in our tests:

https://github.com/bloomberg/memray/blob/960d4c9307ecd896b76e7a32e1b8eeae32d48906/tests/integration/multithreaded_extension/testext.cpp#L91

That would trigger an allocation on exit, which will recreate the TLS.