@pablogsal 2023-01-13
Wow, I tested in my Linux box and this is the difference:

## Before
```
❯ hyperfine 'memray run --trace-python-allocators --native -o /dev/null -m test test_list'
Benchmark 1: memray run --trace-python-allocators --native -o /dev/null -m test test_list
  Time (mean ± σ):      5.069 s ±  0.066 s    [User: 5.020 s, System: 0.059 s]
  Range (min … max):    4.956 s …  5.201 s    10 runs
```

## After
```
❯ hyperfine 'memray run --trace-python-allocators --native -o /dev/null -m test test_list'
Benchmark 1: memray run --trace-python-allocators --native -o /dev/null -m test test_list
  Time (mean ± σ):      2.817 s ±  0.408 s    [User: 2.775 s, System: 0.048 s]
  Range (min … max):    2.471 s …  3.518 s    10 runs
```