@pablogsal 2023-01-13
Hummmm, it seems that we need to still solve the tls problem:

```
{
   <insert_a_suppression_name_here>
   Memcheck:Leak
   match-leak-kinds: definite
   fun:_Znwm
   fun:trackAllocation
   fun:_ZN6memray9intercept6vallocEm
   fun:_ZN12_GLOBAL__N_1L15cleanup_handlerEPv
   fun:__nptl_deallocate_tsd
   fun:__nptl_deallocate_tsd
   fun:start_thread
   fun:clone
}
```