@godlygeek 2023-01-14
> A deallocation can allocate.
Ugh, true.

> Hummmm, it seems that we need to still solve the tls problem
I wonder if this would fix it...

```diff
diff --git a/src/memray/_memray/tracking_api.cpp b/src/memray/_memray/tracking_api.cpp
index 34efe0a..b9b7f52 100644
--- a/src/memray/_memray/tracking_api.cpp
+++ b/src/memray/_memray/tracking_api.cpp
@@ -1144,7 +1144,7 @@ forget_python_stack()
         return;
     }

-    RecursionGuard guard;
+    RecursionGuard::isActive = true;
     PythonStackTracker::get().clear();
 }

```