@pablogsal 2023-01-14
> I wonder if this would fix it...

Nah, the problem seems to be that we are re-creating the vector after the vector was destroyed on thread destruction. This can happen in any thread at any time while the tracker is active so I am not sure how setting the guard on `forget_python_stack` helps