@godlygeek 2023-01-14
Mm. `forget_python_stack` is called when our profile function is uninstalled for a thread, which happens when the Python threadstate for that thread is destroyed, which happens before the thread dies. My thought was that setting the recursion guard there would let us ignore allocations that happen afterwards in that thread.

But that's wrong, anyway - we can create a new threadstate for a thread even after the last one was destroyed (that's normal for a native thread that's occasionally grabbing the GIL to make Python calls). So... bad idea all around.