@pablogsal 2023-01-14
Seems that the monster I commited using `pthread_key` works