@godlygeek 2023-01-14
Eh, OK: I see two ways to solve this problem. Both of them are solutions we applied to Python stack tracking at one point or another.

1. Spend an extra byte in TLS storing a bool representing whether the thread is dying (set it from the destructor of a TLS object). If that bool is set, don't re-create the TLS vector (and instead just supply a stack vector to the NativeTrace constructor for native unwinding).
2. Leveraging our current solution for Python stack tracking: create the vector of native frames when the first Python frame for a thread is entered, and destroy it when that frame is exited. That costs us no extra space in TLS, but it means that we only benefit from caching on threads that have Python stacks, so pure native threads would be slow.

Option 1 seems like a good solution to me, and shouldn't be tough to implement.