@pablogsal 2023-01-14
> Leveraging our current solution for Python stack tracking: create the vector of native frames when the first Python frame for a thread is entered, and destroy it when that frame is exited. That costs us no extra space in TLS, but it means that we only benefit from caching on threads that have Python stacks, so pure native threads would be slow.

But we want this vector to be filled even if a thread never saw any Python frame whatsoever, no?