@pablogsal 2023-01-14
> * Spend an extra byte in TLS storing a bool representing whether the thread is dying (set it from the destructor of a TLS object). If that bool is set, don't re-create the TLS vector (and instead just supply a stack vector to the NativeTrace constructor for native unwinding).

I think this may look uglier than the pthread thing, but I am not sure.