@godlygeek 2023-01-14
> But we want this vector to be filled even if a thread never saw any Python frame whatsoever, no?

Option 2 means we'd get the caching benefit only for Python threads and not for native threads, yeah. Not ideal, but that probably gets us 90% of the benefit of this change. But option 1 gets us 100% of the benefit, at the cost of only 1 more byte of TLS, so that seems better.

> Seems that the monster I commited using `pthread_key` works

I don't understand why... unless... are you sure that you didn't just change the order, so that `key_destructor` is running _after_ `valloc_on_thread_exit`? I'm pretty sure that's what happened:

```text
tests/integration/test_extensions.py In cleanup_handler
in key_destructor
.

=== 1 passed, 448 deselected in 0.73s ===
```