@pablogsal 2023-01-14
> Option 2 means we'd get the caching benefit only for Python threads and not for native threads, yeah. Not ideal, but that probably gets us 90% of the benefit of this change. But option 1 gets us 100% of the benefit, at the cost of only 1 more byte of TLS, so that seems better.
> 
> >

I think I have implemented a naive version of this idea correctly in the last commit, but it uses two new TLSs