@pablogsal 2023-01-14
Ah, and also doesn't work:

```
==6807== 24 bytes in 1 blocks are definitely lost in loss record 117 of 10,110
==6807==    at 0x4849013: operator new(unsigned long) (in /usr/libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
==6807==    by 0x9BE43A6: trackAllocation (tracking_api.h:233)
==6807==    by 0x9BE43A6: memray::intercept::valloc(unsigned long) (hooks.cpp:254)
==6807==    by 0xB34F474: (anonymous namespace)::cleanup_handler(void*) (testext.cpp:61)
==6807==    by 0x4CAA710: __nptl_deallocate_tsd (nptl_deallocate_tsd.c:73)
==6807==    by 0x4CAA710: __nptl_deallocate_tsd (nptl_deallocate_tsd.c:22)
==6807==    by 0x4CAD9C9: start_thread (pthread_create.c:453)
==6807==    by 0x4D3EBB3: clone (clone.S:100)
```