@godlygeek 2023-01-14
It mostly works - it fails if the _first_ call to `trackAllocation` in some thread happens after that thread has already started to exit. As long as something touches the TLS variable before the thread starts to exit, everything is hunky dory, but if the first access happens while TLS is being torn down, then the guard doesn't work: the destructor wasn't scheduled to run at thread teardown because the constructor never ran for that thread.

Ugh.